+++ freecad/MeshStudy/services/run_service.py
@@ -1,9 +1,12 @@
+import FreeCAD as App
 import json
 import os
 from PySide import QtCore
+from freecad.MeshStudy.core.limits import check_mesh_limits
 from freecad.MeshStudy.strategies.registry import get_qoi_extractor, get_refinement_strategy
 from freecad.MeshStudy.core.exceptions import MeshStudyError, MeshError
 from ..__init__ import BACKUP_PATH, DATA_DIR
+from freecad.MeshStudy.objects.mesh_study import check_study_parameters, create_study_results
 
 class MeshStudyRunService:
     """The main excutive file, and the absloute coordinator"""
@@ -20,19 +23,13 @@ class MeshStudyRunService:
         if not hasattr(obj, "Proxy") or not type(obj.Proxy).__name__ == "MeshStudyProxy":
             raise MeshStudyError("Selected object is not a MeshStudy.")   
 
-        # Check the links
-        if not self.obj.TheStudyTarget:
-            raise MeshStudyError("No Analysis target selected in MeshStudy object.")
-    
-        mesh_obj = self.obj.MeshObject
-        solver_obj = self.obj.SolverObject
-        if not mesh_obj or not solver_obj:
-            raise MeshStudyError("Mesh object or Solver object is missing from the analysis.")
-
-        # Check the Mesh size
-        if obj.InitialMeshSize <= 0.0:
-            raise MeshStudyError("Initial Mesh Size must be greater than zero. Please configure the study parameters.")
-
+        # Check the study parameters
+        para_errors = check_study_parameters(obj)
+        if para_errors:
+            for err in para_errors:
+                App.Console.PrintError(f"{err}\n")
+            raise MeshStudyError("Invalid parameters for the MeshStudy object.")
+        
         # get the Refinement method
         qoi_extractor = get_qoi_extractor(self.obj.QuantityOfInterest)
         refinement_strat = get_refinement_strategy()
@@ -43,7 +40,8 @@ class MeshStudyRunService:
             self.obj.NumberOfRuns, 
             self.obj.RefinementFactor
         )
-
+        mesh_obj = obj.MeshObject
+        solver_obj = obj.SolverObject
         results = []
         
         # Repeate along the runs
@@ -62,6 +60,11 @@ class MeshStudyRunService:
             # Check Elementes and nodes number
             nodes = len(mesh_obj.FemMesh.Nodes)
             elements = len(mesh_obj.FemMesh.Volumes)
+            if check_mesh_limits(elements, nodes):
+                if results:
+                    create_study_results(results)
+                    self.clear_results()
+                raise MeshStudyError(f"Run {run_idx}: Mesh limits exceeded. Nodes: {nodes}, Elements: {elements}.")
 
             # Waiting intervals
             if progress_callback:
@@ -89,6 +92,10 @@ class MeshStudyRunService:
             # Results extraction
             result_obj = self.doc.getObject("CCX_Results") or self.doc.getObject(f"CCX_Results_{solver_obj.Name}")
             qoi_value = qoi_extractor.extract(result_obj) if result_obj else 0.0
+
+            # check for Zero QoI
+            if qoi_value == 0.0:
+                raise MeshStudyError(f"Run {run_idx}: Quantity of Interest (QoI) is zero. Check the solver results.")
 
             # format data
             run_data = {